Move most Summary components into one file - make the Change url a prop
